fix(scheduled-messages): satisfy state/lint hardening guard on counters
fire_count INTEGER -> BIGINT (recurring definitions grow it unbounded);
retry_count keeps int4 with an allow-int4 rationale (bounded by the
claim-time MAX_FIRE_RETRIES re-arm cap). Checksum updated for the
not-yet-shipped 0079 migration.

+++ src/db/scheduled_messages.rs
@@ -50,7 +50,7 @@ pub struct ScheduledMessageRow {
     pub expires_at: Option<DateTime<Utc>>,
     pub status: String,
     pub in_flight_delivery_id: Option<String>,
-    pub fire_count: i32,
+    pub fire_count: i64,
     pub last_fired_at: Option<DateTime<Utc>>,
     pub last_error: Option<String>,
     pub source: String,
